requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_dynamodb
aws_cdk.aws_apigatewayv2
aws_cdk.aws_codebuild
aws_cdk.aws_s3
aws_cdk.aws_cloudfront